Desktop/projects/PROJECT 1/agent-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
langgraph>=0.2.28
langchain>=0.3.1
langchain-groq>=0.1.0
langchain-community>=0.3.1
duckduckgo-search>=6.0.0
python-dotenv>=1.0.1
httpx>=0.27.2
sse-starlette>=2.1.3
pydantic>=2.9.2
tenacity>=8.2.3
beautifulsoup4>=4.12.0
requests>=2.31.0
